Switch zmq library
The library I was using had some weird problems with sending messages.
When I was sending with dozens of parallel goroutines, sometimes the
messages would get corrupted. The bytes would come out random in length
and content.

This library does not seem to have that problem, so it is now the
blessed library for use.