Parallelize the agent's workers
This enables us to run N workers in parallel on the agent. The goal here
is to allow workers to handle things that might be slow without gumming
up the works for other clients.

This is rudimentary, as it is fixed at 10 workers. It might be nice to
allow this to adjust automatically, but I'm not entirely sure how I feel
about that.