First pass at a configuration system
Singularity is designed to be able to serve as a monitoring system and
general cluster management tool. The configuration system is our way to
do that.

In short, you can define config items that say "on X, do Y" and then
Singularity will keep track of that. These can be chained, allowing
interesting patterns.

This is currently not yet working, but is in progress.