Fix hostname change alarm
I'm not sure if this is the ideal behavior, but as of this patch, the
hostname change alarm fires if the hostname that the machine has for
itself changes.

This has no bearing on what you decided to call the machine with the
hostname flag. You can call it 'foo' and then your machine named
'ee8f3fe' changes name, the agent will self-destruct. The logic here
is that if the machine's name is changing, probably other stuff is
happening and we should be paranoid and die.