Fix twice-in-one-second possibility
I was noticing warnings in the logs about hosts not updating their lock,
but that was impossible. I could see the updates.

It turns out that the global function can run twice in under a second based
on the randon sleep interval. This fix constrains it to 5-15 seconds and
also makes it so it only warns if something is behind by more than a few
seconds.