Add serial and line based outputs
This adds more output options. Line based output mode is used by default
when multiple hosts are selected, and it prefaces the output with the
name of the host that is returning it. This mode is also line buffered
and prints a line at a time so you don't get interleaved outputs.

The original, and default if you're talking to a single host, is binary
mode. In this mode the output is not adjusted at all. This is more
suitable to, f.ex., catting a remote file.

Finally, serial mode causes the client to wait until a command has
finished before printing output. This will help make sure that command
output isn't interleaved between machines.