Clarify some docs, add -w option
The -w option demotes the "returned non-zero status code" message from
an error log line. This means that it won't show up in standard output.
This is useful sometimes if you are, f.ex., using -A and know that your
command will fail on some nodes and you don't care.