ZMQ simplification+workers patch
This is a big patch. Previously, Singularity used a more complicated
system of sockets in ZMQ. We had ten worker threads, each with their own
socket, and a ZMQ device that pushed messages between them.

We needed a way to determine if the client was still alive, so the agent
had to start sending StillAlive messages down to the client. This means
the client has to respond to them, and the agent needs to track this.

Unfortunately, that didn't work with the way we were designed. A single
worker thread not only blocked when it got a command, but it also wasn't
possible to send a command from the same client to the same worker.
Fixing that could have been done in many ways, but I decided to fix it
by moving to an anonymous, unlimited worker model.

The ZMQ model simplified drastically: now we just have a ROUTER socket
and receive packets. We use the address to allocate a "worker" and then
make sure that any messages that come in with that address are routed to
that worker and that worker only.

That, plus making the worker message pump fully async means that now
workers can send and receive multiple commands and have many things
going on at the same time.

This isn't particularly useful for commands, but it does allow us to
implement the StillAlive messaging. Now the agent knows if a client is
still around and can kill off a job if the client goes away.

This patch is a big change and therefore introduces uncertainty. If I
wasn't the only person on this project and using it, here is where I'd
land this in a testing branch and have people test it. :-)